# Conflicts:
#	Podfile
#	Podfile.lock
#	Pods/Manifest.lock
#	Pods/Pods.xcodeproj/project.pbxproj
#	Pods/Target Support Files/AFNetworking/AFNetworking.xcconfig
#	Pods/Target Support Files/CocoaAsyncSocket/CocoaAsyncSocket.xcconfig
#	Pods/Target Support Files/KVNProgress/KVNProgress.xcconfig
#	Pods/Target Support Files/MJExtension/MJExtension.xcconfig
#	Pods/Target Support Files/Masonry/Masonry.xcconfig
#	Pods/Target Support Files/Pods/Pods.debug.xcconfig
#	Pods/Target Support Files/Pods/Pods.release.xcconfig
#	Pods/Target Support Files/SDWebImage/SDWebImage.xcconfig

+++ HealthyCertificate/ProjectClass/Controller/Appointment/CloudAppointmentCompanyViewController.m
@@ -393,7 +393,7 @@ typedef NS_ENUM(NSInteger, TEXTFILEDTAG)
     }
     _brContract.unitCode = gCompanyInfo.cUnitCode;
     _brContract.unitName = gCompanyInfo.cUnitName;
-    int customercount;
+    NSInteger customercount;
     @try {
        customercount = [_exminationCountField.text integerValue];
     }
